+++ packages/mapbox-react/src/context.ts
@@ -3,14 +3,53 @@ import {
   useContext,
   RefObject,
   useRef,
-  useReducer,
-  useCallback,
-  Reducer,
+  useState,
+  useMemo,
 } from "react";
 import update from "immutability-helper";
 import { Map } from "mapbox-gl";
 import h from "@macrostrat/hyper";
 import { MapPosition } from "@macrostrat/mapbox-utils";
+import { createStore, useStore } from "zustand";
+
+const MapStoreContext = createContext(null);
+
+export function MapboxMapProvider({ children }) {
+  const ref = useRef<Map | null>(null);
+  const [store] = useState(() => {
+    return createStore<MapState>((set) => {
+      return {
+        status: defaultMapStatus,
+        position: null,
+        // Hold a reference to the map object in state
+        ref,
+        dispatch: (action: MapAction): void => {
+          if (action.type === "set-map") {
+            ref.current = action.payload;
+          }
+          set((state) => mapReducer(state, action));
+        },
+      };
+    });
+  });
+
+  return h(MapStoreContext.Provider, { value: store }, children);
+}
+
+function useMapStore<T>(selector: (state: MapState) => T): T {
+  const store = useContext(MapStoreContext);
+  if (!store) {
+    throw new Error("Missing MapStoreProvider");
+  }
+  return useStore(store, selector);
+}
+
+interface MapState {
+  status: MapStatus;
+  position: MapPosition;
+  ref: RefObject<Map | null>;
+  dispatch(action: MapAction): void;
+}
 
 interface MapStatus {
   isLoading: boolean;
@@ -29,33 +68,41 @@ const defaultMapStatus: MapStatus = {
   isStyleLoaded: false,
 };
 
-const MapDispatchContext = createContext<React.Dispatch<MapAction>>(null);
-const MapRefContext = createContext<RefObject<Map | null>>(null);
-const MapStatusContext = createContext<MapStatus>(defaultMapStatus);
-const MapPositionContext = createContext<MapPosition>(null);
-
 export function useMapRef() {
-  return useContext(MapRefContext);
+  return useMapStore((state) => state.ref);
 }
 
-export function useMapStatus() {
-  return useContext(MapStatusContext);
+export function useMapStatus(
+  selector: (state: MapStatus) => any | null = null
+) {
+  return useMapStore(useSubSelector("status", selector));
+}
+
+function useSubSelector(
+  key: string,
+  selector: (state: any) => any | null
+): (state: MapState) => any {
+  return useMemo(() => {
+    if (selector == null) {
+      return (state: MapState) => state[key];
+    } else {
+      return (state: MapState) => selector(state[key]);
+    }
+  }, [selector]);
 }
 
 export function useMapPosition() {
-  return useContext(MapPositionContext);
+  return useMapStore((state) => state.position);
 }
 
 export function useMapElement(): Map | null {
   return useMapRef().current;
 }
 
-export function useMap(): Map | null {
-  return useMapRef().current;
-}
+export const useMap = useMapElement;
 
 export function useMapDispatch() {
-  return useContext(MapDispatchContext);
+  return useMapStore((state) => state.dispatch);
 }
 
 type MapAction =
@@ -65,7 +112,7 @@ type MapAction =
   | { type: "map-moved"; payload: MapPosition }
   | { type: "set-map"; payload: Map };
 
-function mapReducer(state: MapCtx, action: MapAction): MapCtx {
+function mapReducer(state: MapState, action: MapAction): MapCtx {
   switch (action.type) {
     case "set-map":
       return update(state, {
@@ -86,36 +133,4 @@ function mapReducer(state: MapCtx, action: MapAction): MapCtx {
     case "map-moved":
       return { ...state, position: action.payload };
   }
-}
-
-export function MapboxMapProvider({ children }) {
-  const mapRef = useRef<Map | null>();
-  const [value, _dispatch] = useReducer<Reducer<MapCtx, MapAction>>(
-    mapReducer,
-    {
-      status: defaultMapStatus,
-      position: null,
-    }
-  );
-
-  const dispatch = useCallback((action: MapAction) => {
-    if (action.type === "set-map") {
-      mapRef.current = action.payload;
-    }
-    _dispatch(action);
-  }, []);
-
-  return h(
-    MapDispatchContext.Provider,
-    { value: dispatch },
-    h(
-      MapRefContext.Provider,
-      { value: mapRef },
-      h(
-        MapStatusContext.Provider,
-        { value: value.status },
-        h(MapPositionContext.Provider, { value: value.position }, children)
-      )
-    )
-  );
 }