+++ packages/mapbox-react/CHANGELOG.md
@@ -4,13 +4,16 @@ All notable changes to this project will be documented in this file. The format
 is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this
 project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.4.0] - 2024-11-16
+
+- Improve state management using a `zustand` store
+
 ## [2.3.0] - 2024-11-05
 
 - Improve the internal design of the `useMapEaseTo` hook
 - Add some stories for testing
 - Added deprecation warnings to `useMapEaseToCenter` and `useMapEaseToBounds`
 - Add a `useBasicStylePair` hook for getting a basemap in dark or light mode
-
 
 ## [2.2.3] - 2024-10-24